requirements.txt
pandas
numpy
scikit-learn
opencv-python
scikit-image
requests
Pillow
pyarrow